requirements.txt
streamlit
pillow
pymupdf
reportlab

